requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langgraph==0.0.26
langchain==0.1.5
langchain-openai==0.0.5
pydantic==2.5.3
openai==1.10.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
pytz==2024.1
pytest==7.4.4
pillow==10.2.0
